afstandsensor/requirements.txt
neopixel~=0.0.1
rpi.gpio~=0.7.1
board~=1.0.0